Move relationship related model to Model/Relationship
