+++ NOTICE
@@ -1,2 +1,2 @@
 phramework/jsonapi
-Copyright 2015 Xenofon Spafaridis
+Copyright 2015 - 2016 Xenofon Spafaridis
