+++ README.md
@@ -25,7 +25,7 @@ composer test
 
 License
 =======
-Copyright 2015 Xenofon Spafaridis
+Copyright 2015 - 2016 Xenofon Spafaridis
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
