+++ src/Resource.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright 2015 Xenofon Spafaridis
+ * Copyright 2015 - 2016 Xenofon Spafaridis
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,5 +23,32 @@ namespace Phramework\JSONAPI;
  */
 class Resource
 {
+    /**
+     * @var string
+     */
+    public $type;
 
+    /**
+     * @var string
+     */
+    public $id;
+
+    /**
+     * @var object
+     */
+    public $attributes;
+
+    /**
+     * @var object
+     */
+    public $relationships;
+
+    public function __construct($type, $id)
+    {
+        $this->type = $type;
+        $this->id   = $id;
+
+        $this->attributes    = new \stdClass();
+        $this->relationships = new \stdClass();
+    }
 }
